Change persistent services
