--- reactive/src/main/java/com/springRaft/reactive/persistence/log/LogStateRepository.java
@@ -1,5 +1,0 @@
-package com.springRaft.reactive.persistence.log;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-public interface LogStateRepository extends ReactiveCrudRepository<LogState,Long> {}
