--- reactive/src/main/java/com/springRaft/reactive/persistence/state/StateRepository.java
@@ -1,5 +1,0 @@
-package com.springRaft.reactive.persistence.state;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-public interface StateRepository extends ReactiveCrudRepository<State,Long> {}
